fix(i18n): Profile role label uses ROLE_I18N_KEY, helper moved to lib/
The Profile page rendered ``{user.role}`` raw in two places (subtitle
under the name + the "Role" row of "Account Details") with a
``capitalize`` CSS class. That ships the literal API enum value
(``student``, ``pending_teacher``, ``teacher``, ``admin``) to the user
regardless of locale — RU users saw ``Student`` instead of ``Студент``,
and English users saw the snake-cased ``pending_teacher`` with the
underscore visible.

``ROLE_I18N_KEY`` already mapped each ``UserRole`` to the right i18n
key (``roles.student`` / ``roles.pendingTeacher`` / etc.), but lived in
``pages/Admin/dashboard/constants.ts`` — a Profile-page import from
under ``Admin/`` would have been a layering violation.

Move both ``ROLE_I18N_KEY`` and ``ROLE_BADGE_VARIANT`` to a new
``lib/roles.ts`` (single source of truth for role display, sits next
to the ``UserRole`` consumers). Re-export from the old admin constants
file so all four existing call sites (``UsersCard``,
``VirtualAdminUsers``, ``useAdminOverview``) keep working without any
churn.

Profile now uses ``t(ROLE_I18N_KEY[user.role])`` in both spots; the
``capitalize`` utility class is dropped (the translated strings come
out already correctly cased).

EN + RU keysets already had ``roles.*`` — no JSON changes needed.
112 frontend tests green; eslint + tsc clean.
